fixed bug that caused some stairs to not show up, added more floors

+++ app/src/main/java/in/project/enroute/data/model/StairLine.kt
@@ -15,5 +15,5 @@ data class StairLine(
     @SerializedName("stair_polygon_id")
     val stairPolygonId: Int,
     @SerializedName("floors_connected")
-    val floorsConnected: List<Int> = emptyList()
+    val floorsConnected: List<Float> = emptyList()
 )
